Basic model and gameplay for the bedroom and bathroom has been implemented
Currently the user picks an item by simply tapping on it, however the model is general enough such that dragging and releasing something could be easily implemented.

The basic mechanics of the game has been implemented but it has not truly been made into a game yet.

Things to Implement
-timer to limit the amount of time the user has to select the items
-start screen art
-presenting a tips view modally
-supporting sounds and other pretty UI elements
-dragging a cleaning utensil to the mess rather than just tapping it
-enforcement of win loss conditions and all segues related to such

*Important*
-the method currently used to determine if an item has been tapped by the user does not work for all of the items
The method works by looking at the point at which the user tapped and checking to see if the image has a non-zero alpha value there
For some reason this does not work for some of the items.
This could be fixed by editing each asset to have a box of small but non-zero alpha values around the point of interest